// packages/tooling/milestone/types.ts
export type BuildMilestonePackageOptions = {
  url?: string | undefined;
  selector?: string | undefined;
  projectId?: string | undefined;
  width?: number | undefined;
  height?: number | undefined;
  outputBaseDir?: string | undefined;
  sourceFileRelative?: string | undefined;
  sourceFileAbsolute?: string | undefined;
  styleCssAbsolute?: string | undefined;
  studyPngAbsolute?: string | undefined;
};